% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFunction.R
\name{MyFunction}
\alias{MyFunction}
\title{A Simple Function}
\usage{
MyFunction(test_my_package = TRUE)
}
\arguments{
\item{test_my_package}{Do you want to test the package? Defaults to TRUE.}
}
\description{
This function allows you to see how testing and building a package works.
}
\examples{
MyFunction()
}
\keyword{cats}
